Refactored http specific methods from abstract-html-scraper into abstract-scraper
